Enemyのコードをリファクタリング。 Player気合のタスキ問題を修正。 Enemyのルートがしっかり設定されていなかったのを修正。

+++ Sample_05_XX/Sample_05_XX/SrcFile/GameObj/GamePlayer.cpp
@@ -14,7 +14,7 @@
 #include "PlayerState/PlayerDashState.h"
 #include "PlayerState/PlayerReloadState.h"
 
-const float	GamePlayer::BASE_PARAM::PLAYER_MAX_HP = 100.0f;
+const int	GamePlayer::BASE_PARAM::PLAYER_MAX_HP = 100.0f;
 const int	GamePlayer::BASE_PARAM::REGENE_VALUE_SECOND = 50;
 const float GamePlayer::BASE_PARAM::REGENE_COOL_TIME = 5.0f;
 const float GamePlayer::BASE_PARAM::JUMPFORSE = 520.0f;
